client/python/pika-tutorial/requirements.txt
devtools
pika
python-dotenv